+++ packages/entities/entities-certificates/src/components/CACertificateList.vue
@@ -330,7 +330,7 @@ const {
   fetcher,
   fetcherState,
   fetcherCacheKey,
-} = useFetcher({ ...props.config, cacheIdentifier: props.cacheIdentifier }, fetcherBaseUrl.value)
+} = useFetcher(computed(() => ({ ...props.config, cacheIdentifier: props.cacheIdentifier })), fetcherBaseUrl)
 
 const clearFilter = (): void => {
   filterQuery.value = ''