+++ packages/entities/entities-consumer-groups/src/components/ConsumerGroupList.vue
@@ -405,7 +405,7 @@ const {
   fetcher,
   fetcherState,
   fetcherCacheKey,
-} = useFetcher({ ...props.config, cacheIdentifier: props.cacheIdentifier }, fetcherBaseUrl.value, dataKeyName)
+} = useFetcher(computed(() => ({ ...props.config, cacheIdentifier: props.cacheIdentifier })), fetcherBaseUrl, dataKeyName)
 
 const clearFilter = (): void => {
   filterQuery.value = ''