+++ packages/entities/entities-consumers/src/components/ConsumerList.vue
@@ -394,7 +394,7 @@ const {
   fetcher,
   fetcherState,
   fetcherCacheKey,
-} = useFetcher({ ...props.config, cacheIdentifier: props.cacheIdentifier }, fetcherBaseUrl.value, dataKeyName.value)
+} = useFetcher(computed(() => ({ ...props.config, cacheIdentifier: props.cacheIdentifier })), fetcherBaseUrl, dataKeyName)
 
 const clearFilter = (): void => {
   filterQuery.value = ''