+++ packages/entities/entities-plugins/src/components/free-form/shared/types.ts
@@ -1,4 +1,4 @@
-import type { ComponentPublicInstance } from 'vue'
+import type { ComponentPublicInstance, Ref } from 'vue'
 
 type ComponentPublicInstanceConstructor = {
   new (...args: any[]): ComponentPublicInstance<any>
@@ -23,3 +23,84 @@ export type ResetLabelPathRule =
   | 'reset-children' // Inherit paths on its own, but descendant paths start from current path.
   | 'isolate' // Isolate paths for both current and descendants.
   | 'isolate-children' // Inherit paths on its own, children do not inherit.
+// TODO: check lua_schema type
+export interface Field {
+  label: string
+  model: string
+  type?: string
+  default?: any
+  disabled?: boolean
+  help?: string
+  inputType?: string
+  order?: number
+  required?: boolean
+  valueType?: string
+}
+
+export interface Redis {
+  host?: string | null
+  port?: number | null
+  connect_timeout?: number | null
+  send_timeout?: number | null
+  read_timeout?: number | null
+  /**
+   * @description referenceable: true
+   */
+  username?: string | null
+  /**
+   * @description referenceable: true
+   */
+  password?: string | null
+  /**
+   * @description referenceable: true
+   */
+  sentinel_username?: string | null
+  /**
+   * @description referenceable: true
+   */
+  sentinel_password?: string | null
+  database?: number | null
+  keepalive_pool_size?: number | null
+  keepalive_backlog?: number | null
+  sentinel_master?: string | null
+  sentinel_role?: 'master' | 'slave' | 'any' | null
+  sentinel_nodes?: RedisSentinelNode[] | null // `"len_min": 1`
+  cluster_nodes?: RedisClusterNode[] | null // `"len_min": 1`
+  ssl?: boolean
+  ssl_verify?: boolean
+  server_name?: string | null
+  cluster_max_redirections?: number | null
+  connection_is_proxied?: boolean
+}
+
+export type RedisPartialType = 'redis-ce' | 'redis-ee'
+
+export type RedisTypeDisplay = 'Host/Port' | 'Sentinel' | 'Cluster'
+
+export interface RedisConfig {
+  name: string
+  type: RedisPartialType
+  config: Redis
+}
+
+export interface FlattendRedisConfigurationFields extends Redis {
+  name: string
+  created_at: string
+  updated_at: string
+  type: RedisPartialType
+}
+export interface RedisSentinelNode {
+  host: string
+  port?: number | null
+}
+
+export interface RedisClusterNode {
+  ip: string
+  port?: number | null
+}
+
+export type PartialInfo = {
+  redisType: Ref<RedisPartialType | undefined>
+  redisPath: Ref<string | undefined>
+  isEditing: boolean
+}