+++ packages/entities/entities-routes/src/types/index.ts
@@ -2,6 +2,5 @@
 // The actual types and interfaces should be contained in separate files within this folder.
 
 export * from './route-list'
-export * from './method-badge'
 export * from './route-form'
 export * from './route-config-card'