+++ packages/entities/entities-upstreams-targets/src/components/TargetsList.vue
@@ -259,7 +259,7 @@ const {
   fetcher,
   fetcherState,
   fetcherCacheKey,
-} = useFetcher({ ...props.config, cacheIdentifier: props.cacheIdentifier }, fetcherBaseUrl.value)
+} = useFetcher(computed(() => ({ ...props.config, cacheIdentifier: props.cacheIdentifier })), fetcherBaseUrl)
 
 const resetPagination = (): void => {
   // Increment the cache key on sort