+++ packages/core/expressions/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.39](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.38...@kong-ui-public/expressions@0.9.39) (2025-05-21)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.9.38](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.37...@kong-ui-public/expressions@0.9.38) (2025-05-21)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.9.37](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.36...@kong-ui-public/expressions@0.9.37) (2025-05-19)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.9.36](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.35...@kong-ui-public/expressions@0.9.36) (2025-05-14)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
 ## [0.9.35](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.34...@kong-ui-public/expressions@0.9.35) (2025-05-06)
 
 **Note:** Version bump only for package @kong-ui-public/expressions