+++ packages/entities/entities-keys/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.8.27](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.8.26...@kong-ui-public/entities-keys@3.8.27) (2025-05-21)
+
+**Note:** Version bump only for package @kong-ui-public/entities-keys
+
+
+
+
+
+## [3.8.26](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.8.25...@kong-ui-public/entities-keys@3.8.26) (2025-05-19)
+
+**Note:** Version bump only for package @kong-ui-public/entities-keys
+
+
+
+
+
+## [3.8.25](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.8.24...@kong-ui-public/entities-keys@3.8.25) (2025-05-14)
+
+
+### Bug Fixes
+
+* **entities:** fetcher url reactivity ([#2120](https://github.com/Kong/public-ui-components/issues/2120)) ([16cbc20](https://github.com/Kong/public-ui-components/commit/16cbc20058a3acda085ebc18b16fb034104126c9))
+
+
+
+
+
 ## [3.8.24](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.8.23...@kong-ui-public/entities-keys@3.8.24) (2025-05-06)
 
 **Note:** Version bump only for package @kong-ui-public/entities-keys